+++ packages/CodingStandard/tests/Fixer/Commenting/RemoveUselessDocBlockFixer/ConfiguredTest.php
@@ -3,7 +3,7 @@
 namespace Symplify\CodingStandard\Tests\Fixer\Commenting\RemoveUselessDocBlockFixer;
 
 use Iterator;
-use Symplify\EasyCodingStandard\Testing\AbstractCheckerTestCase;
+use Symplify\EasyCodingStandardTester\Testing\AbstractCheckerTestCase;
 
 /**
  * @see \Symplify\CodingStandard\Fixer\Commenting\RemoveUselessDocBlockFixer