+++ packages/CodingStandard/tests/Fixer/ControlStructure/RequireFollowedByAbsolutePathFixer/RequireFollowedByAbsolutePathFixerTest.php
@@ -2,7 +2,7 @@
 
 namespace Symplify\CodingStandard\Tests\Fixer\ControlStructure\RequireFollowedByAbsolutePathFixer;
 
-use Symplify\EasyCodingStandard\Testing\AbstractCheckerTestCase;
+use Symplify\EasyCodingStandardTester\Testing\AbstractCheckerTestCase;
 
 /**
  * @see \Symplify\CodingStandard\Fixer\ControlStructure\RequireFollowedByAbsolutePathFixer