+++ packages/CodingStandard/tests/Sniffs/Architecture/ExplicitExceptionSniff/ExplicitExceptionSniffTest.php
@@ -3,7 +3,7 @@
 namespace Symplify\CodingStandard\Tests\Sniffs\Architecture\ExplicitExceptionSniff;
 
 use Iterator;
-use Symplify\EasyCodingStandard\Testing\AbstractCheckerTestCase;
+use Symplify\EasyCodingStandardTester\Testing\AbstractCheckerTestCase;
 
 /**
  * @see \Symplify\CodingStandard\Sniffs\Architecture\ExplicitExceptionSniff